requirements.txt
alembic==1.6.5
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
clickclick==20.10.2
connexion==2.9.0
Flask==1.1.4
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.21
swagger-ui-bundle==0.0.8
toml==0.10.2
urllib3==1.26.6
Werkzeug==1.0.1
